lispat_app/requirements.txt
docx2txt
nltk
pdfminer.six
pygogo
PyPDF2
python-docx
textract
tabula-py
